requirements.txt
Flask==2.1.1
Flask-Cors==3.1.1
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==2.5.1
paramiko==2.9.2
python-dotenv==0.19.2
